Enhanced version of calculator.py library.
The Calmath library includes most of the math calculations in just one file. The same is true for calculator.py's program, but it has added square root, circle area, triangle area, rectangle area, and more. To learn more about the library, scroll down to the bottom of the program and click the link.